Added Statistics and Daily Summary

+++ database/migrations/2020_06_06_235529_create_wheat_stocks_table.php
@@ -24,7 +24,7 @@ class CreateWheatStocksTable extends Migration
             $table->integer('num_of_sack')->unsigned();
             $table->integer('weight_per_sack')->unsigned();
             $table->integer('price')->unsigned();
-            $table->integer('commission')->unsigned();
+            $table->float('commission')->unsigned();
             $table->string('category');
             $table->string('status')->nullable()->default('unpaid');
