+++ database/migrations/2020_06_08_070927_create_rice_stocks_table.php
@@ -24,7 +24,7 @@ class CreateRiceStocksTable extends Migration
             $table->integer('num_of_sack')->unsigned();
             $table->integer('weight_per_sack')->unsigned();
             $table->integer('price')->unsigned();
-            $table->integer('commission')->unsigned();
+            $table->float('commission')->unsigned();
             $table->bigInteger('rice_type_id')->unsigned();
             $table->foreign('rice_type_id')->references('id')->on('rice_types');
             $table->string('category');
